+++ src/main/java/shop/gaship/gashipscheduler/domain/gradehistory/dto/request/GradeHistoryAddRequestDto.java
@@ -1,4 +1,4 @@
-package shop.gaship.gashipscheduler.gradehistory.dto.request;
+package shop.gaship.gashipscheduler.domain.gradehistory.dto.request;
 
 import lombok.AllArgsConstructor;
