+++ src/main/java/shop/gaship/gashipscheduler/domain/member/adapter/MemberAdapter.java
@@ -1,9 +1,9 @@
-package shop.gaship.gashipscheduler.member.adapter;
+package shop.gaship.gashipscheduler.domain.member.adapter;
 
 import java.time.LocalDate;
 import reactor.core.publisher.Flux;
-import shop.gaship.gashipscheduler.member.dto.request.MemberModifyRequestDto;
-import shop.gaship.gashipscheduler.member.dto.response.MemberResponseDto;
+import shop.gaship.gashipscheduler.domain.member.dto.request.MemberModifyRequestDto;
+import shop.gaship.gashipscheduler.domain.member.dto.response.MemberResponseDto;
 
 /**
  * shopping mall rest-api 에 회원과 관련한 요청을 보내는 adapter.
